test(update): replace mockist checks with user journeys
